+++ CHANGES.rst
@@ -7,6 +7,8 @@ Bug Fixes
 Cubeviz
 ^^^^^^^
 
+- Re-enable support for exporting spectrum-viewer. [#2825]
+
 Imviz
 ^^^^^